ml_project/requirements.txt
numpy==1.23.4
pandas==1.5.1
seaborn==0.12.1
matplotlib==3.6.0
click==8.1.3
PyYAML==6.0
scikit-learn==1.1.2
marshmallow-dataclass==8.5.9
pytest==7.2.0
flake8==5.0.4